Drop forty no-op 32-bit casts that restate a field's own declared type
Same class as the byte-width pass: (f32) on f32 fields (Hcurves plane
normals, camera fov/height/distance, EffectBox extents, WaveAnimato
amplitudes, GroundAnima radius/yOffset/maxSinkDepth, KT_Torch progress,
DIMLavaBall firePeriod, DIM2IceFloe curveStep, WM_spiritpl heightOffset,
607/437 phase, textrender window width), (int)/(s32) on int fields (249
gameBitId, 250 hitPriority, 599's yaw counters) and (u32) on
obj->objectFlags. All twenty-two objects byte-identical.

Reverted from the same sweep because they move bytes or change meaning:
276 (int)state->step, 329 (int)obj->anim.alpha, audio.c, plus two the
detector mis-typed (578's placementData is short*, objhits's is void*).

+++ src/dlls/engine/20_Hcurves/Hcurves.c
@@ -136,7 +136,7 @@ u8 gObjfsaWalkGroupActive[0xB8];
     }                                                                                                                  \
     pl->normalX = (s16)(32767.0f * dxn);                                                                \
     pl->normalZ = (s16)(32767.0f * dzn);                                                                \
-    *(po) = -((f32)pl->normalX * (XA) + (f32)pl->normalZ * (ZA))
+    *(po) = -(pl->normalX * (XA) + pl->normalZ * (ZA))
 
 static inline f32 RomCurveNode_GetHermiteTangent(void** nodePtr, int angleOffset, int useCos);
 inline f32 objfsaCorner(s8 ofs, f32 scl, f32* base);
